feat: consolidate comment tests
